Fix for returnto Url with Autologin enabled
When using PluggableAuth and wgPluggableAuth_EnableAutoLogin the
doBeforeInitialize hook issues a HTML redirect to the Userlogin page.
This was done with Mediawiki OutputPage redirects, which forgets the
parameter for the returnto site.
It was replaced by bare PHP header('Location: ...');
Also the returnto url was urlencode one time to much, so all returnto
links to urls with namespaces and to images/media redirected to Mainpage

Change-Id: I51de765b08ac91a1c49e6ca6fa5db01f2efbf96b