Enable logout button if locallogin is on.
If both AutoLogin is enabled, and LocalLogin in on, then the
functionality when arriving to the wiki is that the user is sent to the
login page.  Because they can login via wiki credentials in this case,
they also need to be able to log out, meaning the logout button
shouldn't be removed.

Change-Id: I4658def872ee8f5c396c6b3e3be74bb9a7d6bcc4