Allow PluggableAuth to work with multiple authentication plugin instances
This allows PluggableAuth to be configured with more than one
authentication plugin or with more than one instance of the same
authentication plugin (i.e. different identity providers
with the same authentication plugin) simultaneously.

Configuration is provided through the $wgPluggableAuth_Config
configuration variable or, if that is not set, from the legacy
configuration variables for backward compatibility
($wgPluggableAuth_Class, $wgPluggableAuth_ButtonLabelMessage,
$wgPluggableAuth_ButtonLabel, and $wgPluggableAuth_ExtraLoginFields).
If $wgPluggableAuth_Config is set, it must be an array of arrays
that must contain array fields as follows:

- class (required): the name of a PluggableAuth subclass that will be
used to instantiate the authentication plugin
- data (optional): will be passed to the constructor of the
authentication plugin, if set; if no 'data' field is provided, no
arguments will be passed to the constructor
- buttonLabelMessage (optional): a Message that will be used for the
login button label
- buttonLabel (optional): a text string that will be used for the
login button label if buttonLabelMessage is not set
- extraLoginFields (optional): an array of fields to be added to the
login form (see documentation at AuthenticationRequest::getFieldInfo()
for the format). If not set, the value will come from the static
getExtraLoginFields() function on the authentication plugin. That
function defaults to an empty array (no extra login fields) in the
PluggableAuth abstract superclass.

Bug: T258726
Change-Id: Id11008cd280ba2bcac68a8c1ecf00b86b1c49b77