Redirect to login page with an error message if no returnToUrl is found
This means that the original return-to-URL is lost for some reason.  This can happen if the user appears with an ancient
session cookie or, as in T373409, they end up on the Special:PluggableAuthLogin without a session.

Bug: T373409
Change-Id: I4d2e0e93a496bf264dc76905ff13426b7e823f8f
(cherry picked from commit 2de9f07a5f5100dd37dbee7614af795037f4cfe5)